Fix mypy issues
